+++ examples/01-load-and-save.lua
@@ -1,15 +1,10 @@
--- We can load models from a file with remodel.load
-local game = remodel.load("test-models/folder-and-value.rbxmx")
-
--- game is a DataModel -- remodel.load always wraps its return value in a
--- DataModel instance because models and places can contain multiple top-level
--- instances.
-print("game is a ", game.ClassName)
+-- We can load models from a file with remodel.readModelFile.
+-- Model files can contain multiple top-level instances so it returns a list.
+local root = remodel.readModelFile("test-models/folder-and-value.rbxmx")[1]
 
 -- Remodel is a subset of Roblox's API with stricter semantics.
 -- You should feel right at home if you work on Roblox!
-local Root = game:FindFirstChild("Root")
-local stringValue = Root:FindFirstChild("String")
+local stringValue = root:FindFirstChild("String")
 
 print("Found string value: ", stringValue.Name)