+++ examples/03-arguments.lua
@@ -1,4 +1,5 @@
 -- You can pass arguments to the script by passing extra values to remodel!
--- Try running this example as `remodel examples/03-arguments.lua hello`
+-- Try running this example as:
+-- remodel examples/03-arguments.lua "hello, world"
 
 print("My arguments:", ...)