+++ test-scripts/read-place.lua
@@ -1,0 +1,6 @@
+local game = remodel.readPlaceFile("test-models/place-with-models.rbxlx")
+
+assert(game.ClassName == "DataModel")
+
+assert(game.ReplicatedStorage.Name == "ReplicatedStorage")
+assert(game.Workspace.Baseplate.Name == "Baseplate")